requirements.txt
fastapi
sqlalchemy[asyncio]
aiomysql
alembic
greenlet
uvicorn
cryptography
pymysql
python-dotenv
itsdangerous
python-multipart
flask
flask-login
flask-mail
werkzeug
mysql-connector-python
jinja2
python-jose[cryptography]
redis
fastapi-limiter
sqladmin
itsdangerous
pyrate_limiter
